requirements.txt
Flask
feedparser
requests
python-dateutil
beautifulsoup4